+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "BusinessMonitor",
-  "version": "0.1.0+209",
+  "version": "0.1.0+210",
   "private": true,
   "type": "module",
   "engines": {
